Update qasm tests
